// src/util/enums/enum.ts
export enum StageEnum {
    dev = 'dev',
    prod = 'prod'
}


export enum NameEnum {
    LandingPage = "Landing Page",
    BootcampList = "Bootcamp List",
    BootcampDetail = "Bootcamp Details",
    Signup = "Sign Up",
    Login = "Login",
    User = "User",
    AdminUsers = "AdminUsers",
    AdminBootcamps = "AdminBootcamps",
    AdminBootcamp = "AdminBootcamp",
    Publisher = "Publisher",
}